API

QInsControlCore.Controller — Type

Controller(instrnm, addr)

construct a Controller to send commands to the instrument determined by (instrnm, addr). A Controller has three ways to send commands to the Processor: wirte read query.

julia> ct(write, cpu, "*IDN?", Val(:write))
"done"

julia> ct(read, cpu, Val(:read))
"read"

julia> ct(query, cpu, "*IDN?", Val(:query))
"query"

the commands needed to execute is not necessary to be wirte, read or query.

julia> idn_get(instr) = query(instr, "*IDN?")
idn_get (generic function with 1 method)

julia> ct(idn_get, cpu, Val(:read))
"query"

the definition of function idn_get must happen before the Controller ct is logged in or one can log out and log in again.

QInsControlCore.Processor — Type

Processor()

construct a Processor to deal with the commands sended into by Controllers.

Base.read — Method

read(instr)

read the instrument.

Base.write — Method

write(instr, msg)

write some message string to the instrument.

QInsControlCore.connect! — Method

connect!(rm, instr)

connect to an instrument with given ResourceManager rm.

connect!(instr)

same but with auto-generated ResourceManager.

QInsControlCore.disconnect! — Method

disconnect!(instr)

disconnect the instrument.

QInsControlCore.fast! — Method

fast!(cpu::Processor)

change the cpu mode to fast mode. Default mode is slow mode. The fast mode is not necessary in most cases.

QInsControlCore.find_resources — Method

find_resources(cpu::Processor)

auto-detect available instruments.

QInsControlCore.instrument — Method

instrument(name, addr)

generate an instrument with (name, addr) which automatically determines the type of this instrument.

QInsControlCore.isconnected — Method

isconnected(instr)

determine if the instrument is connected.

QInsControlCore.login! — Method

login!(cpu::Processor, ct::Controller)

log the Controller in the Processor which can be done before and after the cpu started.

QInsControlCore.logout! — Method

logout!(cpu::Processor, ct::Controller)

log the Controller out the Processor.

logout!(cpu::Processor, addr::String)

log all the Controllers that control the instrument with address addr out the Processor.

QInsControlCore.query — Method

query(instr, msg; delay=0)

query the instrument with some message string.

QInsControlCore.reconnect! — Method

reconnect!(cpu::Processor)

reconnect the instruments that log in the Processor.

QInsControlCore.slow! — Method

slow!(cpu::Processor)

change the cpu mode to slow mode. Default mode is slow mode, which decrease the cpu cost.

QInsControlCore.start! — Method

start!(cpu::Processor)

start the Processor.

QInsControlCore.stop! — Method

stop!(cpu::Processor)

stop the Processor.